requirements.txt
requests==2.28.1
jupyterlab==3.4.7
pyspark==3.2.1
PyYAML==6.0
ipykernel==6.15.0
pandas
pymongo[srv]
python-dotenv
apache-airflow
-e .